requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
depcheck==0.3.2
deprecation==2.1.0
dj-database-url==3.0.1
Django==5.2.4
django-ckeditor==6.7.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-js-asset==3.1.2
django-modeltranslation==0.19.16
django-solo==2.4.0
django-storages==1.14.6
django-unfold==0.63.0
djangorestframework==3.16.0
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
pydeps==1.12.20
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
realtime==2.5.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stdlib-list==0.11.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
websockets==15.0.1
whitenoise==6.9.0
